DIE now performs complete reload on idb change. Closes #3.
Before tihs, DIE was never fully unloaded. This caused several issues:
1. The DIEdb did not clear when changing the IDB you are working on.
2. Menu initialization failed after first time DIE was loaded
3. When modifying DIE, a complete reload of IDA was required.

As DIE is only relevant when a file is loaded, it was changed to be a PLUGIN_PROC plugin.
Also, menu cleanups were added to the plugin termination.
To solve the DIEdb issue, the global `__die_db` is re-initialized inside `DieManager`. This is a bad fix (better to change it so that it is not global at all), but it works.